desktop: Move --timedemo stuff to time_demo.rs
